Improved compatibility with IE 7 and 8, removed minified script.
Older versions of internet explorer do not understand
how to handle .apply with arguements that are not
defined, so we have to pass only this as an argument
when we know there are no additional parameters to
pass.

Minification should either be done with a build
script or not at all. Opting for not at all for
now.